requirements.txt
google-ai-generativelanguage==0.6.6
streamlit==1.36.0
python-dotenv==1.0.1